Create shared LessonDialog component with overflow-y-auto and interactive quizzes
- New LessonDialog.tsx with proper scrolling and clickable quiz answers
- Updated LearningLessonCard.tsx to use shared component (Dashboard)
- Updated ScratchPadExperience.tsx to use shared component (Orbit)
- Fixes scroll issue on Dashboard and makes quizzes interactive on both pages
