+++ public/sitemap-0.xml
@@ -1,18 +1,18 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>https://toodl.co/about</loc><lastmod>2026-01-02T03:25:48.836Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://toodl.co/faq</loc><lastmod>2026-01-02T03:25:48.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://toodl.co/data-deletion</loc><lastmod>2026-01-02T03:25:48.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://toodl.co/scratch-pad</loc><lastmod>2026-01-02T03:25:48.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://toodl.co/privacy</loc><lastmod>2026-01-02T03:25:48.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://toodl.co/journal</loc><lastmod>2026-01-02T03:25:48.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://toodl.co/dashboard</loc><lastmod>2026-01-02T03:25:48.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://toodl.co/terms</loc><lastmod>2026-01-02T03:25:48.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://toodl.co/feedback</loc><lastmod>2026-01-02T03:25:48.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://toodl.co/settings</loc><lastmod>2026-01-02T03:25:48.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://toodl.co/orbit</loc><lastmod>2026-01-02T03:25:48.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://toodl.co/flow</loc><lastmod>2026-01-02T03:25:48.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://toodl.co/quest</loc><lastmod>2026-01-02T03:25:48.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://toodl.co/budget</loc><lastmod>2026-01-02T03:25:48.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://toodl.co/split</loc><lastmod>2026-01-02T03:25:48.837Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://toodl.co/about</loc><lastmod>2026-01-02T03:53:11.254Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://toodl.co/data-deletion</loc><lastmod>2026-01-02T03:53:11.255Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://toodl.co/feedback</loc><lastmod>2026-01-02T03:53:11.255Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://toodl.co/faq</loc><lastmod>2026-01-02T03:53:11.255Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://toodl.co/split</loc><lastmod>2026-01-02T03:53:11.255Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://toodl.co/privacy</loc><lastmod>2026-01-02T03:53:11.255Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://toodl.co/budget</loc><lastmod>2026-01-02T03:53:11.255Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://toodl.co/flow</loc><lastmod>2026-01-02T03:53:11.255Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://toodl.co/terms</loc><lastmod>2026-01-02T03:53:11.255Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://toodl.co/settings</loc><lastmod>2026-01-02T03:53:11.255Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://toodl.co/scratch-pad</loc><lastmod>2026-01-02T03:53:11.255Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://toodl.co/dashboard</loc><lastmod>2026-01-02T03:53:11.255Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://toodl.co/orbit</loc><lastmod>2026-01-02T03:53:11.255Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://toodl.co/quest</loc><lastmod>2026-01-02T03:53:11.255Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://toodl.co/journal</loc><lastmod>2026-01-02T03:53:11.255Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
 </urlset>
