+++ src/components/orbit/LearningLessonCard.tsx
@@ -3,10 +3,9 @@
 import { useState } from "react";
 import { Button } from "@/components/ui/button";
 import { Card, CardHeader, CardTitle } from "@/components/ui/card";
-import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@/components/ui/dialog";
-import { ScrollArea } from "@/components/ui/scroll-area";
 import { cn } from "@/lib/utils";
 import type { OrbitLearningLesson } from "@/types/orbit";
+import { LessonDialog } from "./LessonDialog";
 
 export function LearningLessonCard({
     lesson,
@@ -21,21 +20,8 @@ export function LearningLessonCard({
 }) {
     const tone = isNight ? "text-indigo-200" : "text-slate-600";
     const [open, setOpen] = useState(false);
-    const quizItems = lesson.quiz ?? [];
     const [saving, setSaving] = useState(false);
     const [saveMessage, setSaveMessage] = useState<string | null>(null);
-    const [quizResponses, setQuizResponses] = useState<
-        Record<number, { selected: string; isCorrect: boolean }>
-    >({});
-
-    const handleAnswerSelect = (questionIndex: number, answer: string, correctAnswer: string) => {
-        const isCorrect =
-            answer.trim().toLowerCase() === (correctAnswer ?? "").trim().toLowerCase();
-        setQuizResponses((prev) => ({
-            ...prev,
-            [questionIndex]: { selected: answer, isCorrect },
-        }));
-    };
 
     const handleSaveLesson = async () => {
         if (!userId) {
@@ -69,6 +55,7 @@ export function LearningLessonCard({
             setSaving(false);
         }
     };
+
     return (
         <Card
             className={cn(
@@ -131,107 +118,13 @@ export function LearningLessonCard({
                     {saving ? "Saving..." : "Save for Later"}
                 </Button>
             </div>
-            <Dialog open={open} onOpenChange={setOpen}>
-                <DialogContent
-                    className={cn(
-                        "max-w-3xl max-h-[85vh] border-none p-0 flex flex-col overflow-hidden",
-                        isNight ? "bg-slate-900/90 text-white" : "bg-white/95 text-slate-900"
-                    )}
-                >
-                    <DialogHeader className="px-4 sm:px-6 pt-4 sm:pt-6 pb-2">
-                        <DialogTitle className={cn("text-lg sm:text-xl font-semibold", isNight ? "text-white" : "text-slate-900")}>
-                            {lesson.title}
-                        </DialogTitle>
-                        <p className={cn("text-xs sm:text-sm", tone)}>
-                            Day {lesson.day} of {lesson.totalDays} · {lesson.overview}
-                        </p>
-                    </DialogHeader>
-                    <ScrollArea className="flex-1 min-h-0">
-                        <div className="px-4 sm:px-6 pb-4 sm:pb-6 space-y-3 text-sm leading-relaxed">
-                            {lesson.paragraphs.map((paragraph, index) => (
-                                <p key={index} className={isNight ? "text-indigo-100" : "text-slate-700"}>
-                                    {paragraph}
-                                </p>
-                            ))}
-                            {lesson.code && lesson.code.length ? (
-                                <div className="space-y-2">
-                                    {lesson.code.map((block, codeIdx) => (
-                                        <pre
-                                            key={codeIdx}
-                                            className={cn(
-                                                "overflow-auto rounded-xl border px-4 py-3 text-xs font-mono",
-                                                isNight
-                                                    ? "border-white/10 bg-slate-900/80 text-emerald-100"
-                                                    : "border-slate-200 bg-slate-50 text-emerald-700"
-                                            )}
-                                        >
-                                            {block}
-                                        </pre>
-                                    ))}
-                                </div>
-                            ) : null}
-                        </div>
-                        {quizItems.length > 0 && (
-                            <div className="mt-6 space-y-3 rounded-2xl border border-dashed border-emerald-200/60 p-4">
-                                <p className={cn("text-xs font-semibold uppercase tracking-[0.35em]", tone)}>Quick quiz</p>
-                                {quizItems.map((item, idx) => (
-                                    <div key={idx} className="space-y-2 rounded-xl border border-white/10 bg-white/5 p-3">
-                                        <p className={cn("text-sm font-semibold", isNight ? "text-white" : "text-slate-900")}>
-                                            {item.question}
-                                        </p>
-                                        <div className="space-y-2 text-sm">
-                                            {item.answers.map((answer, answerIdx) => {
-                                                const response = quizResponses[idx];
-                                                const selected = response?.selected === answer;
-                                                const isCorrect = response?.isCorrect ?? false;
-                                                const correctAnswer = item.correctAnswer ?? "";
-                                                const isRightAnswer =
-                                                    (answer ?? "").trim().toLowerCase() === correctAnswer.trim().toLowerCase();
-                                                return (
-                                                    <button
-                                                        key={answerIdx}
-                                                        type="button"
-                                                        onClick={() => handleAnswerSelect(idx, answer, correctAnswer)}
-                                                        className={cn(
-                                                            "w-full rounded-lg border px-3 py-2 text-left transition",
-                                                            isNight
-                                                                ? "border-white/15 text-indigo-100 hover:border-emerald-300/60"
-                                                                : "border-slate-200 text-slate-700 hover:border-emerald-300/60",
-                                                            selected && isCorrect && (isNight ? "border-emerald-300/80 bg-emerald-500/10" : "border-emerald-400 bg-emerald-50"),
-                                                            selected && !isCorrect && (isNight ? "border-red-300/80 bg-red-500/10" : "border-red-300 bg-red-50"),
-                                                            !selected && response && isRightAnswer && (isNight ? "border-emerald-200/60" : "border-emerald-300")
-                                                        )}
-                                                    >
-                                                        {answer}
-                                                    </button>
-                                                );
-                                            })}
-                                            {quizResponses[idx] ? (
-                                                <p
-                                                    className={cn(
-                                                        "text-xs font-semibold",
-                                                        quizResponses[idx].isCorrect
-                                                            ? isNight
-                                                                ? "text-emerald-200"
-                                                                : "text-emerald-700"
-                                                            : isNight
-                                                                ? "text-red-200"
-                                                                : "text-red-700"
-                                                    )}
-                                                >
-                                                    {quizResponses[idx].isCorrect
-                                                        ? "Correct!"
-                                                        : `Not quite. Correct answer: ${item.correctAnswer}`}
-                                                </p>
-                                            ) : null}
-                                        </div>
-                                    </div>
-                                ))}
-                            </div>
-                        )}
-                    </ScrollArea>
-                </DialogContent>
-            </Dialog>
+
+            <LessonDialog
+                lesson={lesson}
+                open={open}
+                onOpenChange={setOpen}
+                isNight={isNight}
+            />
         </Card>
     );
 }
